Adds user login and logout functionality
Uses the DRF token authentication classes to
create and store a token for login for a particular
user. For that I am using the drf built-in function.

For logout we require the user to be authenticated, then we
can access and delete that user's token. When they try to do
anything else that requires authentication the will be
unable to until they hit the login point again.