Adds permission for isMemberofSub to update posts
I have been working on this permission for a while. I just
got past the bug regarding the browsable api checking every
possible request type for permissions that was causing issues.

It has now occured to me that this may be something that is handled
better with a serializer validator rather than a view permission.
I think I will implement it like that. First I want to create a
through table for hte user <-> sub_membership relationship. It is
obviously going to be important going forward and it would be nice
to be able to refer to a through model like that directly and handle
things a little more explicity.