Adds sub membership table to users
This is handled now and everything can save. In the sub
serializer the members and moderators m2m fields are now
read only and they are set in the view to the user creating
the sub.

Later I will add a new view for the original
moderator to add additional moderators.