Adds populate script
This script uses Faker and the requests library to
access the api and create users and subs. I am using
it so that I have a little more content
to work on while styling the front end. I also need to get some
critical mass of posts/subs so that I can get some kind of
popular/new/hot sorting working.

I'm not sure this is the best way to go. It would be nicer to
pull some content from the actual reddit api, at least
subreddit names and descriptions. I may do that in the future but
for now this seems reasonable.

I have hit a natural stopping point now that I am trying to
subscribe some of the fake users to the fake subreddits. I have
not implemented a subscribe api endpoint yet. I will do that next
and then keep working on this.