Adds sub/unsub endpoint for subreddits
An autorized user can now subscribe and unsubscribe. I came
across an issue in the userSubMembership through table
that it was allowing duplicate memberships, i.e. the same
user to the same sub. I have migrated to include a
unique together on those two fields/columns in the through table.