Adds comment tree view and serializer
This is a detail view for a particular comment. It calls
the mptt.get_descendants() function and builds a json
tree to be returned out of all of the descendants.

The next step is to create a view that can
return a tree for all root comments of a particular post.