Refactors the way that comments are related to posts and parents
Now requires a parent_fn 'full name' of the style 'tx_pk' where
x is either 1 or 2 to indicate whether it is a child or root comment
respectively. And pk is the primary key for the 'parent'.

t1_pk indicates that the comment is the child of another comment
so the post field will be set to null and the parent field will
be set to the comment inicated by pk.

Similarly t2_pk indicates that the comment is a root comment on
the post with supplied pk. In that case the parent will be set to
null and post will be set to the indicated post.

This avoids needing the consumer to provide a post and a parent for
each comment creation. It was becoming a headache to need to
provide a post when commenting on another comment that already
knew the post.

In the case of t1_pk it might be better to still set the post. I'm not
sure. I'm leaving it blank now because it doesn't seem necessary and
I think it makes indicating root comments easier.