Moves upvotes from a db field to a simple attr of the comment model
I did this because we now have the CommentVote m2m field
and table. We don't need to keep a separate count of the current
vote. For a particular comment we can just do an aggregation on
the reverse queryset (i.e. the queryset from CommentVote with
the comment of interest involved) and sum the 'vote_type' column.