Catches error when vote_type is not provided
This achievment is a side-effect of what I had been working on
but gave up. I wanted the api consumer to provide
vote_type as 'upvote/downvote/no_vote' rather that '1/-1/0'.
I got that to work but it was a little messy. I also
didn't want to spend any more time making the reply feature
the appropriate form. For some reason the serializer.field
'source' argument wasn't working as I expected.