Creates vote app and moves CommentVote there
I think that it is worth it, and will make things much simpler
for consumers, to create a votes app that handles both
post and comment votes. Now we can again take advantage of the
'full name' structure for the id's of posts and comments
within the new vote view.

This commit just made the app and moved the pre-existing comment
vote functionality there. The next step is to make it work for posts
as well.