Modifies vote_state for comment creation and retreival
Upon comment creation a vote_state of 0 is set. In the future
it may be set to one and a vote will be created on behalf of the
creator, this mimics what reddit does.

Prviously when retreiving the tree if the auth user
had not voted on a comment then the vote_state was set to
null. I think that is inconsistent and it was causing
some problems in the front-end. I changed it so that
if there is not vote yet from this user it will return
vote_state=0.