Moves to pk related fields for post serializer
I am moving away from the hyperlink related fields. They are not
really of much use for me now (they are nice in the browsable
api though). I also changed the field name of sub to subreddit
in both the model and the serializer. This is more consistent with
what I am calling things.