Adds comment delete functionality
We are going with the reddit flavor of deletion. The
comment itself is not removed from the db but
some of the fields are set to indicate it's deletion.

Foremost is the deleted field that I just added. I am envisioning
that being useful in the frontend. The other option was to
rely on the poster being null or not to indicate a deleted comment.
I can't think of a reason why a non-deleted comment would have
a null poster but I think just making the explicit deleted field
is probably worth it for simplicity.