REBASE this one
I am in the middle of adding popular/home subreddit functionality.
I stupidly built everything in the sub detail view. In retrospect
that might be a good place for it actually. But as it stands now
I need to move it to SubPostListView which takes a subreddit
name in the GET request URL and retuns the appropriate postList.