Adds handling for popular and home subreddits
I decided for the moment to not make them true subreddits.
Mostly because then I wouldn't be able to use a FK from
post to subreddit, since the posts would need to be
members of home and potentiall popular in addition
to their own true subreddit.

I'm not totally satisfied with the setup. It's a little
messy and it doesn't give me an opportunity to work with
celery. For now it will work.