Refactors subreddit subscription view and serializer
Previously we just used the SubSerializer and did a lot
of work in the subscription view that was really most
appropriately placed in a customr serializer. This commit
creates that serializer and updates/simplifies/renames the view.

I also finally begain with some testing. I added tests for
subreddit subscription.