Adds test for subreddit creation and detail view
This gets a pretty good chunk of the subreddit testing
completed. One outstaning area is post creation. The more
I look at how it is right now the less I like the way I did it.
I'm not sure yet whether to test it in the sub app or the post app.

The functionality is in this app because post creation is necessarily
tied to a specific subreddit. There is also a lot of logic in
the view and I wonder if it might be possible to move that to a serializer.
And maybe move the whole thing to the post app.