Moves post creation from the sub app to the post app
I was originally located in the subreddit app because I was
trying to mimic the reddit design where you can go to a
subreddit url + /post and create a post.

I decided that it would make a lot more sense to just have
it in the post app for general cohesion.

I also created some test for post creation that are included
in this commit.