Adds a user profile detail view and serializer
This is in contrast to the user detail view which is
related to user authentication, i.e. when a user is logged
in then they may see their own user detail view. In retrospect
I should have named that one userAuth or something.

This new view returns the user's username, posts and comments. The
idea is that it can be used to populate a use profile page.