Change user profile serializer fields comments and posts
Instead of using nested serializers we are now using a
serializerMethodField so that we can handle the ordering
of comments and posts on the UserProfileSerializer itself.
We want them in order from newest to oldest at the moment
but we can more easily update that in the future to include
'hot' or whatever.