Refactor of votes view to change how request.data is accessed
I was modifying request.data directly up until this point. When I run
tests on this app the requests.data becomes immutable. I also
discovered that I can't assign to request, though that probably was the
case all along.

I changed how the serializer is selected to accomadate this
newfound immutability. It now works with the normal api stuff
as well as the tests.