Improves design of vote assignment in post serializer
Previously we had a GET query parameter username that was read
to determine how to set the vote state for each post in the response.

That was a poor design because it wasn't very clean but more imporantly
it meant that anyone could use anyone elses username to see
their votes. This commit changes the post serializer to check
if there is an authenticated user and then look up post
votes based on that.