Refactor post view for retrieving posts associated with a subreddit
I previuosly had a bunch of custom get methods for each pseudo
subreddit. That was going to make it pretty difficult to include
pagination. Instead I am moving that functionality to the get_queryset
function that is called by the list funciton prior to pagination.
Now I just need to implement a get_{}_queryset function for each
pseudo subreddit.

This commit includes the creation of the new get_queryset function
and the first of these, the get_home_queryset. It includes tests.